requirements.txt
PyGithub
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
XlsxWriter
python-gitlab
